requirements.txt
alembic==0.9.6
chardet==3.0.4
click==6.7
dataset==1.0.2
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
normality==0.4.4
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
